project/requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
asgiref==3.7.2
asyncpg==0.28.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.3.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
